server/pyproject.toml
[tool.poetry]
name = "thingbox"
version = "0.1.0"
description = ""
authors = ["Barney Mannerings <b.mannerings@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
tweepy = "^3.10.0"
cachetools = "^4.2.2"
PyNaCl = "^1.4.0"
base58 = "^2.1.0"
uvicorn = "^0.15.0"
aiofiles = "^0.7.0"
click = "^8.0.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
